Feat: made request lazy
Made the request properties body, queryParameters and urlParameters lazy
to allow them to be used in the middleware but not extracted if the
pipeline is stop before the are neede.

Example: if auth middleware checks a bearer token and this fails the
body, url and queryParameters are never checked by the extracting
fucntions.