[bug] Stop trying to load ts files as endpoints
Init was trying to load files with a js or ts extention, this is not
logical as modules are loaded after being transpiled into javascript
and we would never load a ts file.

Errors would occur whenever a user generated declration files as these
will end in .d.ts tipi would try to load them as modules and the would
crash the program as the where trying to use import outside a module.